V V

+++ so_long/srcs/srcs_mandatory/create_map.c
@@ -6,11 +6,11 @@
 /*   By: csakamot <csakamot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/23 17:02:15 by csakamot          #+#    #+#             */
-/*   Updated: 2023/08/14 16:52:03 by csakamot         ###   ########.fr       */
+/*   Updated: 2023/09/03 12:51:13 by csakamot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/so_long.h"
+#include "../../includes/so_long.h"
 
 void	ft_put_img(t_game *game, void	*img, int x, int y)
 {
@@ -19,8 +19,8 @@ void	ft_put_img(t_game *game, void	*img, int x, int y)
 
 void	ft_create_map(t_game *game, t_img *texture)
 {
-	int	width;
-	int	height;
+	size_t	width;
+	size_t	height;
 
 	height = 0;
 	while (height < game -> map_h)
